pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "hackathon-project"
version = "0.1.0"
description = "StudyAbroad AI - AI-powered study abroad counsellor"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi==0.109.0",
  "uvicorn[standard]==0.27.0",
  "sqlalchemy==2.0.23",
  "psycopg2-binary==2.9.9",
  "python-dotenv==1.0.0",
  "pydantic==2.5.3",
]

[tool.setuptools]
packages = ["backend"]
include-package-data = true

[tool.setuptools.package-data]
backend = ["**/*.py"]
